requirements.txt
lask
opencv-python
numpy
tensorflow
matplotlib
Pillow